requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
requests
nicegui
python-jose[cryptography]
passlib[bcrypt]
alembic
google-generativeai